[BEAM-2366] Move generate_proto_files into setup.py
