+++ sdks/python/setup.py
@@ -20,8 +20,14 @@
 from __future__ import absolute_import
 from __future__ import print_function
 
+import contextlib
+import glob
+import inspect
 import os
 import platform
+import re
+import shutil
+import subprocess
 import sys
 import warnings
 from distutils import log
@@ -35,9 +41,11 @@ from pkg_resources import DistributionNotFound
 from pkg_resources import get_distribution
 from pkg_resources import normalize_path
 from pkg_resources import to_filename
+from pkg_resources import resource_filename
 from setuptools import Command
 from setuptools.command.build_py import build_py
 from setuptools.command.sdist import sdist
+
 
 class mypy(Command):
   user_options = []
@@ -214,24 +222,268 @@ AWS_REQUIREMENTS = [
     'boto3 >=1.9'
 ]
 
+BEAM_PROTO_PATHS = [
+    os.path.join('..', '..', 'model', 'pipeline', 'src', 'main', 'proto'),
+    os.path.join('..', '..', 'model', 'job-management', 'src', 'main', 'proto'),
+    os.path.join('..', '..', 'model', 'fn-execution', 'src', 'main', 'proto'),
+]
 
-# We must generate protos after setup_requires are installed.
-def generate_protos_first(original_cmd):
+PYTHON_PROTO_OUTPUT_PATH = os.path.join('apache_beam', 'portability', 'api')
+
+MODEL_RESOURCES = [
+    os.path.normpath('../../model/fn-execution/src/main/resources'\
+            + '/org/apache/beam/model/fnexecution/v1/standard_coders.yaml'),
+]
+
+
+def generate_urn_files(log, out_dir):
+  """
+  Create python files with statically defined URN constants.
+
+  Creates a <proto>_pb2_urn.py file for each <proto>_pb2.py file that contains
+  an enum type.
+
+  This works by importing each api.<proto>_pb2 module created by `protoc`,
+  inspecting the module's contents, and generating a new side-car urn module.
+  This is executed at build time rather than dynamically on import to ensure
+  that it is compatible with static type checkers like mypy.
+  """
+  import google.protobuf.message as message
+  import google.protobuf.pyext._message as pyext_message
+
+  class Context(object):
+    INDENT = '  '
+    CAP_SPLIT = re.compile('([A-Z][^A-Z]*|^[a-z]+)')
+
+    def __init__(self, indent=0):
+      self.lines = []
+      self.imports = set()
+      self.empty_types = set()
+      self._indent = indent
+
+    @contextlib.contextmanager
+    def indent(self):
+      self._indent += 1
+      yield
+      self._indent -= 1
+
+    def prepend(self, s):
+      if s:
+        self.lines.insert(0, (self.INDENT * self._indent) + s + '\n')
+      else:
+        self.lines.insert(0, '\n')
+
+    def line(self, s):
+      if s:
+        self.lines.append((self.INDENT * self._indent) + s + '\n')
+      else:
+        self.lines.append('\n')
+
+    def import_type(self, typ):
+      modname = typ.__module__
+      if modname in ('__builtin__', 'builtin'):
+        return typ.__name__
+      else:
+        self.imports.add(modname)
+        return modname + '.' + typ.__name__
+
+    @staticmethod
+    def is_message_type(obj):
+      return isinstance(obj, type) and \
+             issubclass(obj, message.Message)
+
+    @staticmethod
+    def is_enum_type(obj):
+      return type(obj).__name__ == 'EnumTypeWrapper'
+
+    def python_repr(self, obj):
+      if isinstance(obj, message.Message):
+        return self.message_repr(obj)
+      elif isinstance(obj, (list,
+                            pyext_message.RepeatedCompositeContainer,  # pylint: disable=c-extension-no-member
+                            pyext_message.RepeatedScalarContainer)):  # pylint: disable=c-extension-no-member
+        return '[%s]' % ', '.join(self.python_repr(x) for x in obj)
+      else:
+        return repr(obj)
+
+    def empty_type(self, typ):
+      name = ('EMPTY_' +
+              '_'.join(x.upper()
+                       for x in self.CAP_SPLIT.findall(typ.__name__)))
+      self.empty_types.add('%s = %s()' % (name, self.import_type(typ)))
+      return name
+
+    def message_repr(self, msg):
+      parts = []
+      for field, value in msg.ListFields():
+        parts.append('%s=%s' % (field.name, self.python_repr(value)))
+      if parts:
+        return '%s(%s)' % (self.import_type(type(msg)), ', '.join(parts))
+      else:
+        return self.empty_type(type(msg))
+
+    def write_enum(self, enum_name, enum, indent):
+      ctx = Context(indent=indent)
+
+      with ctx.indent():
+        for v in enum.DESCRIPTOR.values:
+          extensions = v.GetOptions().Extensions
+
+          prop = (
+              extensions[beam_runner_api_pb2.beam_urn],
+              extensions[beam_runner_api_pb2.beam_constant],
+              extensions[metrics_pb2.monitoring_info_spec],
+              extensions[metrics_pb2.label_props],
+          )
+          reprs = [self.python_repr(x) for x in prop]
+          if all(x == "''" or x.startswith('EMPTY_') for x in reprs):
+            continue
+          ctx.line('%s = PropertiesFromEnumValue(%s)' %
+                   (v.name, ', '.join(self.python_repr(x) for x in prop)))
+
+      if ctx.lines:
+        ctx.prepend('class %s(object):' % enum_name)
+        ctx.prepend('')
+        ctx.line('')
+      return ctx.lines
+
+    def write_message(self, message_name, message, indent=0):
+      ctx = Context(indent=indent)
+
+      with ctx.indent():
+        for obj_name, obj in inspect.getmembers(message):
+          if self.is_message_type(obj):
+            ctx.lines += self.write_message(obj_name, obj, ctx._indent)
+          elif self.is_enum_type(obj):
+            ctx.lines += self.write_enum(obj_name, obj, ctx._indent)
+
+      if ctx.lines:
+        ctx.prepend('class %s(object):' % message_name)
+        ctx.prepend('')
+      return ctx.lines
+
+  pb2_files = [path for path in glob.glob(os.path.join(out_dir, '*_pb2.py'))]
+  api_path = os.path.dirname(pb2_files[0])
+  sys.path.insert(0, os.path.dirname(api_path))
+
+  def _import(m):
+    # TODO: replace with importlib when we drop support for python2.
+    return __import__('api.%s' % m, fromlist=[None])
+
   try:
-    # See https://issues.apache.org/jira/browse/BEAM-2366
-    # pylint: disable=wrong-import-position
-    import gen_protos
+    beam_runner_api_pb2 = _import('beam_runner_api_pb2')
+    metrics_pb2 = _import('metrics_pb2')
 
-    class cmd(original_cmd, object):
-      def run(self):
-        gen_protos.generate_proto_files(log=log)
-        super(cmd, self).run()
-    return cmd
-  except ImportError as err:
-    warnings.warn(
-        "Could not import gen_protos, skipping proto generation: %s (pwd=%s)" %
-        (err, os.getcwd()))
-    return original_cmd
+    for pb2_file in pb2_files:
+      modname = os.path.splitext(pb2_file)[0]
+      out_file = modname + '_urns.py'
+      modname = os.path.basename(modname)
+      mod = _import(modname)
+
+      ctx = Context()
+      for obj_name, obj in inspect.getmembers(mod):
+        if ctx.is_message_type(obj):
+          ctx.lines += ctx.write_message(obj_name, obj)
+
+      if ctx.lines:
+        for line in reversed(sorted(ctx.empty_types)):
+          ctx.prepend(line)
+
+        for modname in reversed(sorted(ctx.imports)):
+          ctx.prepend('from . import %s' % modname)
+
+        ctx.prepend('from ..utils import PropertiesFromEnumValue')
+
+        log.info("Writing urn stubs: %s" % out_file)
+        with open(out_file, 'w') as f:
+          f.writelines(ctx.lines)
+
+  finally:
+    sys.path.pop(0)
+
+
+def generate_proto_files(force=False):
+
+  py_sdk_root = os.path.dirname(os.path.abspath(__file__))
+  proto_dirs = [os.path.join(py_sdk_root, path) for path in BEAM_PROTO_PATHS]
+  proto_files = sum(
+      [glob.glob(os.path.join(d, '*.proto')) for d in proto_dirs], [])
+  out_dir = os.path.join(py_sdk_root, PYTHON_PROTO_OUTPUT_PATH)
+  out_files = [path for path in glob.glob(os.path.join(out_dir, '*_pb2.py'))]
+
+  if out_files and not proto_files and not force:
+    # We have out_files but no protos; assume they're up to date.
+    # This is actually the common case (e.g. installation from an sdist).
+    log.info('No proto files; using existing generated files.')
+    return
+
+  elif not out_files and not proto_files:
+    if not os.path.exists(py_sdk_root):
+      raise RuntimeError(
+          'Not in apache git tree; unable to find proto definitions.')
+    else:
+      raise RuntimeError(
+          'No proto files found in %s.' % proto_dirs)
+
+  if force:
+    regenerate = 'forced'
+  elif not out_files:
+    regenerate = 'no output files'
+  elif len(out_files) < len(proto_files):
+    regenerate = 'not enough output files'
+  elif (
+      min(os.path.getmtime(path) for path in out_files)
+      <= max(os.path.getmtime(path)
+             for path in proto_files + [os.path.realpath(__file__)])):
+    regenerate = 'output files are out-of-date'
+  elif len(out_files) > len(proto_files):
+    regenerate = 'output files without corresponding .proto files'
+    # too many output files: probably due to switching between git branches.
+    # remove them so they don't trigger constant regeneration.
+    for out_file in out_files:
+      os.remove(out_file)
+  else:
+    regenerate = None
+
+  if regenerate:
+    from grpc_tools import protoc
+
+    log.info('Regenerating Python proto definitions (%s).' % regenerate)
+    builtin_protos = resource_filename('grpc_tools', '_proto')
+    args = (
+        [sys.executable] +  # expecting to be called from command line
+        ['--proto_path=%s' % builtin_protos] +
+        ['--proto_path=%s' % d for d in proto_dirs] +
+        ['--python_out=%s' % out_dir] +
+        # TODO(robertwb): Remove the prefix once it's the default.
+        ['--grpc_python_out=grpc_2_0:%s' % out_dir] +
+        proto_files)
+    ret_code = protoc.main(args)
+    if ret_code:
+      raise RuntimeError(
+          'Protoc returned non-zero status (see logs for details): '
+          '%s' % ret_code)
+
+    # copy resource files
+    for path in MODEL_RESOURCES:
+      shutil.copy2(os.path.join(py_sdk_root, path), out_dir)
+
+    ret_code = subprocess.call(
+        ["futurize", "--both-stages", "--write", "--no-diff", out_dir])
+    if ret_code:
+      raise RuntimeError(
+          'Error applying futurize to generated protobuf python files.')
+
+    generate_urn_files(log, out_dir)
+
+# We must generate protos after requirements from pyproject.toml are installed.
+def generate_protos_first(original_cmd):
+
+  class cmd(original_cmd, object):
+    def run(self):
+      generate_proto_files()
+      super(cmd, self).run()
+  return cmd
 
 
 python_requires = '>=2.7,!=3.0.*,!=3.1.*,!=3.2.*,!=3.3.*,!=3.4.*'
